requirements.txt
# Core dependencies
fastapi>=0.68.0
sqlalchemy>=1.4.25
pydantic>=1.8.2
pydantic-settings>=2.0.0 # Added for BaseSettings
bcrypt>=3.2.0
PyJWT>=2.3.0
redis>=4.3.4
python-dotenv>=0.19.0
pytest>=6.2.5

# Development dependencies
# (Add any dev dependencies here if needed)
cryptography>=3.4.7